+++ fliqo-member-api/src/main/java/com/fliqo/exception/MemberException.java
@@ -1,0 +1,14 @@
+package com.fliqo.exception;
+
+public class MemberException extends RuntimeException {
+    private final MemberError error;
+
+    public MemberException(MemberError error) {
+        super(error.message());
+        this.error = error;
+    }
+
+    public MemberError getError() {
+        return error;
+    }
+}
