--- fliqo-member-api/src/main/java/com/fliqo/exception/MemberSignupException.java
@@ -1,14 +1,0 @@
-package com.fliqo.exception;
-
-public class MemberSignupException extends RuntimeException {
-    private final MemberSignupError error;
-
-    public MemberSignupException(MemberSignupError error) {
-        super(error.message());
-        this.error = error;
-    }
-
-    public MemberSignupError getError() {
-        return error;
-    }
-}
