+++ fliqo-member-api/src/main/java/com/fliqo/service/validator/MemberPolicyValidator.java
@@ -5,7 +5,7 @@ import java.util.regex.Pattern;
 import org.springframework.stereotype.Component;
 
 import com.fliqo.domain.repository.MemberRepository;
-import com.fliqo.exception.MemberSignupError;
+import com.fliqo.exception.MemberError;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,17 +17,17 @@ public class MemberPolicyValidator {
     /**
      * 주어진 이메일이 회원가입에 사용 가능한지(중복되지 않는지) 검증합니다.
      *
-     * <p>이미 동일한 이메일을 가진 회원이 존재하면 {@link com.fliqo.exception.MemberSignupError#EMAIL_ALREADY_EXISTS}
-     * 예외를 발생시켜 가입 절차를 중단합니다. 존재하지 않으면 아무 동작 없이 반환됩니다.
+     * <p>이미 동일한 이메일을 가진 회원이 존재하면 {@link MemberError#EMAIL_ALREADY_EXISTS} 예외를 발생시켜 가입 절차를 중단합니다.
+     * 존재하지 않으면 아무 동작 없이 반환됩니다.
      *
      * @param email 중복 여부를 확인할 이메일(사전 검증/정규화된 값)
-     * @throws RuntimeException {@link com.fliqo.exception.MemberSignupError#EMAIL_ALREADY_EXISTS} 에
-     *     해당하는 예외가 발생합니다(이미 사용 중인 경우).
-     * @see com.fliqo.exception.MemberSignupError#EMAIL_ALREADY_EXISTS
+     * @throws RuntimeException {@link MemberError#EMAIL_ALREADY_EXISTS} 에 해당하는 예외가 발생합니다(이미 사용 중인
+     *     경우).
+     * @see MemberError#EMAIL_ALREADY_EXISTS
      */
     public void ensureEmailAvailable(String email) {
         if (memberRepository.existsByEmail(email)) {
-            throw MemberSignupError.EMAIL_ALREADY_EXISTS.get();
+            throw MemberError.EMAIL_ALREADY_EXISTS.get();
         }
     }
 
@@ -45,11 +45,11 @@ public class MemberPolicyValidator {
      */
     public void validateOrThrow(String password, String passwordConfirm) {
         if (!password.equals(passwordConfirm)) {
-            throw MemberSignupError.PASSWORD_MISMATCH.get();
+            throw MemberError.PASSWORD_MISMATCH.get();
         }
 
         if (!PWD.matcher(password).matches()) {
-            throw MemberSignupError.PASSWORD_POLICY_VIOLATION.get();
+            throw MemberError.PASSWORD_POLICY_VIOLATION.get();
         }
     }
 }
